:bath: Combine article tests into canonical and amp
